requirements.txt
faster-whisper 
gtts 
transformers 
llama-cpp-python 
sounddevice 
scipy
